+++ app/form_objects/subscribe_form.rb
@@ -13,7 +13,7 @@ class SubscribeForm
 
   validates :subscribeable_id, :subscribeable_type, :character, :status, presence: true
   validates :subscribeable_type, inclusion: { in: %w[Event Static] }
-  validates :status, inclusion: 0..4
+  validates :status, inclusion: 0..5
   validates :for_role, inclusion: 0..2, allow_nil: true
   validates :comment, length: { maximum: 100 }, allow_nil: true
   validate :subscribeable_exists?
@@ -49,6 +49,7 @@ class SubscribeForm
   def status_to_integer
     return status if status.is_a?(Integer)
     case status
+      when 'declined' then 5
       when 'reserve' then 4
       when 'approved' then 3
       when 'signed' then 2